+++ src/tooling.ts
@@ -10,10 +10,10 @@ export interface ToolDefinition {
 }
 
 export const testedReleaseSet = {
-  id: "toolchain-preview-2026-07-23",
-  url: "https://pawnkit.dev/release-sets/toolchain-preview-2026-07-23.json",
+  id: "build-preview-2026-07-24",
+  url: "https://pawnkit.dev/release-sets/build-preview-2026-07-24.json",
   versions: {
-    pawn: "v1.1.3",
+    pawn: "v1.2.0",
     pawnfmt: "v1.3.5",
     pawnlint: "v1.1.4",
     pawnlsp: "v0.26.2",